:recycle: Move all recipe page components into separate folder
